Consolidate FakeImageLoader impls
Just putting it in src/main/kotlin + `@VisibleForTesting` since test fixtures aren't supported yet for kotlin android proejcts
